drl-trading-ingest/pyproject.toml
[build-system]
requires = ["setuptools>=46.1.0"]
build-backend = "setuptools.build_meta"

[project]
name = "drl-trading-ingest"
description = "Data ingestion package for AI-powered trading system with support for multiple data sources and real-time streaming"
authors = [
    {name = "sonntagnico", email = "nico-tor@hotmail.com"},
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
]
version = "0.1.0"
dependencies = [
    # Core dependencies for runtime
    "numpy",
    "pandas",
    "yfinance", # Yahoo Finance API
    "boto3", # AWS S3 integration
    "pyarrow", # Parquet file support
    "requests", # HTTP requests for APIs
    "websocket-client", # WebSocket connections for real-time data
    # API and services
    "click",
    # Dependency injection framework
    "dependency_injector",
    "pandas-stubs",
    "flask>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "kafka-python>=2.2.13",
    # Database migration
    "alembic>=1.13.0",
    "sqlalchemy>=2.0.0",
]

[project.optional-dependencies]
testing = [
    "pytest",
    "pytest-cov",
    "pytest-benchmark",
]
dev = [
    "mypy",
    "ruff",
    "ipython",
]

[tool.ruff]
target-version = "py38"
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "B",    # flake8-bugbear
    "I",    # isort
]
ignore = ["E501"]  # Line length is handled by black

[tool.ruff.lint.per-file-ignores]
"notebooks/*.ipynb" = ["E402", "F403"]
"tests/**/*" = ["F841"]  # Allow unused variables in tests

[tool.ruff.lint.isort]
known-first-party = ["drl_trading_ingest"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_optional = true
ignore_missing_imports = true
cache_dir = ".mypy_cache"
incremental = true
follow_imports = "silent"
exclude = "(.*tests/.*)|(.*todo.*)\\.py$"

[[tool.mypy.overrides]]
module = ["pandas.*", "numpy.*", "pytest.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = [
    "tests/unit",
    "tests/integration"
]
python_files = "*_test.py *_it.py"
python_classes = "Test*"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s %(levelname)s %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')"
]
addopts = [
    "-ra",
    "-m not slow"
]

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests"]

[tool.coverage.run]
source = ["drl_trading_ingest"]
branch = true
omit = [
    "*/tests/*",
    "*/setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
    "if False:",
]
